Cleanup xbucket step
